+++ CHANGELOG.md
@@ -3,6 +3,7 @@ v0.4.0 (in development)
 - Increased MSRV to 1.81
 - Linux release artifacts are now built on Ubuntu 22.04 (up from Ubuntu 20.04),
   which may result in a more recent glibc being required
+- Added a `--json` option to `query-tracker`
 
 v0.3.1 (2025-02-23)
 -------------------